- Reorganized Folder Structure - added jsconfig for alias absolute file path
