+++ src/components/Layout/Layout.js
@@ -1,0 +1,245 @@
+import React, { useState } from 'react'
+import { Link } from 'react-router-dom'
+import { Typography, CssBaseline, Box, useMediaQuery } from '@mui/material'
+import { createTheme, ThemeProvider, withStyles } from '@material-ui/core/styles';
+
+import NavigationSidebar from 'components/Layout/Sidebar/NavigationSidebar';
+import Header from 'components/Layout/Header/Header'
+import MainContent from 'components/Layout/Content/MainContent'
+
+
+function Copyright() {
+    return (
+        <Typography variant="body2" color="text.secondary" align="center">
+            {'Copyright © '}
+            <Link color="inherit" to="/admin/login">
+                Scrumptious
+            </Link>{' '}
+            {new Date().getFullYear()}.
+        </Typography>
+    );
+}
+
+let theme = createTheme({
+    palette: {
+        primary: {
+            light: '#63ccff',
+            main: '#009be5',
+            dark: '#006db3',
+        },
+    },
+    typography: {
+        h5: {
+            fontWeight: 500,
+            fontSize: 26,
+            letterSpacing: 0.5,
+        },
+    },
+    shape: {
+        borderRadius: 8,
+    },
+    components: {
+        MuiTab: {
+            defaultProps: {
+                disableRipple: true,
+            },
+        },
+    },
+    mixins: {
+        toolbar: {
+            minHeight: 48,
+        },
+    },
+});
+
+theme = {
+    ...theme,
+    components: {
+        MuiDrawer: {
+            styleOverrides: {
+                paper: {
+                    backgroundColor: '#081627',
+                },
+            },
+        },
+        MuiButton: {
+            styleOverrides: {
+                root: {
+                    textTransform: 'none',
+                },
+                contained: {
+                    boxShadow: 'none',
+                    '&:active': {
+                        boxShadow: 'none',
+                    },
+                },
+            },
+        },
+        MuiTabs: {
+            styleOverrides: {
+                root: {
+                    marginLeft: theme.spacing(1),
+                },
+                indicator: {
+                    height: 3,
+                    borderTopLeftRadius: 3,
+                    borderTopRightRadius: 3,
+                    backgroundColor: theme.palette.common.white,
+                },
+            },
+        },
+        MuiTab: {
+            styleOverrides: {
+                root: {
+                    textTransform: 'none',
+                    margin: '0 16px',
+                    minWidth: 0,
+                    padding: 0,
+                    [theme.breakpoints.up('md')]: {
+                        padding: 0,
+                        minWidth: 0,
+                    },
+                },
+            },
+        },
+        MuiIconButton: {
+            styleOverrides: {
+                root: {
+                    padding: theme.spacing(1),
+                },
+            },
+        },
+        MuiTooltip: {
+            styleOverrides: {
+                tooltip: {
+                    borderRadius: 4,
+                },
+            },
+        },
+        MuiDivider: {
+            styleOverrides: {
+                root: {
+                    backgroundColor: 'rgb(255,255,255,0.15)',
+                },
+            },
+        },
+        MuiListItemButton: {
+            styleOverrides: {
+                root: {
+                    '&.Mui-selected': {
+                        color: '#4fc3f7',
+                    },
+                },
+            },
+        },
+        MuiListItemText: {
+            styleOverrides: {
+                primary: {
+                    fontSize: 14,
+                    fontWeight: theme.typography.fontWeightMedium,
+                },
+            },
+        },
+        MuiListItemIcon: {
+            styleOverrides: {
+                root: {
+                    color: 'inherit',
+                    minWidth: 'auto',
+                    marginRight: theme.spacing(2),
+                    '& svg': {
+                        fontSize: 20,
+                    },
+                },
+            },
+        },
+        MuiAvatar: {
+            styleOverrides: {
+                root: {
+                    width: 32,
+                    height: 32,
+                },
+            },
+        },
+    },
+};
+
+const drawerWidth = 256;
+
+const styles = {
+    root: {
+        display: 'flex',
+        minHeight: '100vh',
+    },
+    drawer: {
+        [theme.breakpoints.up('sm')]: {
+            width: drawerWidth,
+            flexShrink: 0,
+        },
+    },
+    app: {
+        flex: 1,
+        display: 'flex',
+        flexDirection: 'column',
+    },
+    main: {
+        flex: 1,
+        padding: theme.spacing(6, 4),
+        background: '#fff',
+    },
+    footer: {
+        padding: theme.spacing(2),
+        background: '#eaeff1',
+    },
+};
+
+
+const Layout = () => {
+
+    const [mobileOpen, setMobileOpen] = useState(false);
+    const isSmUp = useMediaQuery(theme.breakpoints.up('sm'));
+    //const isLoggedIn = useSelector(state => state.logged)
+
+    const handleDrawerToggle = () => {
+        setMobileOpen(!mobileOpen);
+    };
+
+
+
+    return (
+        <ThemeProvider theme={theme}>
+            <Box sx={{ display: 'flex', minHeight: '100vh' }}>
+                <CssBaseline />
+                <Box
+                    component="nav"
+                    sx={{ width: { sm: drawerWidth }, flexShrink: { sm: 0 } }}
+                >
+                    {isSmUp ? null : (
+                        <NavigationSidebar
+                            PaperProps={{ style: { width: drawerWidth } }}
+                            variant="temporary"
+                            open={mobileOpen}
+                            onClose={handleDrawerToggle}
+                        />
+                    )}
+
+                    <NavigationSidebar
+                        PaperProps={{ style: { width: drawerWidth } }}
+                        sx={{ display: { sm: 'block', xs: 'none' } }}
+                    />
+                </Box>
+                <Box sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
+                    <Header onDrawerToggle={handleDrawerToggle} />
+                    <MainContent />
+
+                    <Box component="footer" sx={{ p: 2, bgcolor: '#eaeff1' }}>
+                        <Copyright />
+                    </Box>
+                </Box>
+            </Box>
+        </ThemeProvider>
+    )
+}
+
+
+
+export default withStyles(styles)(Layout)
